Change preload to js
